+++ lib/src/graph/operations/copy.dart
@@ -6,7 +6,7 @@ extension CopyGraphExtension<V, E> on Graph<V, E> {
   /// If [empty] is set to true, returns a new graph of the same
   /// type without the vertices and edges.
   Graph<V, E> copy({bool empty = false}) {
-    final result = Graph<V, E>.create(
+    final result = Graph<V, E>(
       isDirected: isDirected,
       vertexStrategy: vertexStrategy,
     );