+++ test/graph/graph_test.dart
@@ -6,7 +6,7 @@ import '../utils/graph.dart';
 void main() {
   group('directed', () {
     test('empty', () {
-      final graph = Graph<String, int>.create(isDirected: true);
+      final graph = Graph<String, int>(isDirected: true);
       expect(graph.isDirected, isTrue);
       expect(graph.vertices, isEmpty);
       expect(graph.edges, isEmpty);
@@ -14,21 +14,21 @@ void main() {
     });
     group('modifying', () {
       test('add vertex', () {
-        final graph = Graph<String, int>.create(isDirected: true);
+        final graph = Graph<String, int>(isDirected: true);
         graph.addVertex('Hello');
         expect(graph.vertices, ['Hello']);
         expect(graph.edges, isEmpty);
         expectInvariants(graph);
       });
       test('add vertices', () {
-        final graph = Graph<String, int>.create(isDirected: true);
+        final graph = Graph<String, int>(isDirected: true);
         graph.addVertices(['Hello', 'World']);
         expect(graph.vertices, ['Hello', 'World']);
         expect(graph.edges, isEmpty);
         expectInvariants(graph);
       });
       test('remove vertex', () {
-        final graph = Graph<String, int>.create(isDirected: true);
+        final graph = Graph<String, int>(isDirected: true);
         graph.addVertex('Hello');
         graph.removeVertex('Hello');
         expect(graph.vertices, isEmpty);
@@ -36,7 +36,7 @@ void main() {
         expectInvariants(graph);
       });
       test('add edge', () {
-        final graph = Graph<String, int>.create(isDirected: true);
+        final graph = Graph<String, int>(isDirected: true);
         graph.addEdge('Hello', 'World', value: 42);
         expect(graph.vertices, unorderedEquals(['Hello', 'World']));
         expect(graph.edges, [
@@ -46,7 +46,7 @@ void main() {
         expectInvariants(graph);
       });
       test('add self-edge', () {
-        final graph = Graph<String, int>.create(isDirected: true);
+        final graph = Graph<String, int>(isDirected: true);
         graph.addEdge('Myself', 'Myself', value: 42);
         expect(graph.vertices, unorderedEquals(['Myself']));
         expect(
@@ -58,7 +58,7 @@ void main() {
         expectInvariants(graph);
       });
       test('put edge', () {
-        final graph = Graph<String, List<int>>.create(isDirected: true);
+        final graph = Graph<String, List<int>>(isDirected: true);
         graph.putEdge('a', 'b', () => []).add(1);
         graph.putEdge('b', 'a', () => []).add(2);
         expect(graph.vertices, unorderedEquals(['a', 'b']));
@@ -72,7 +72,7 @@ void main() {
         expectInvariants(graph);
       });
       test('remove edge', () {
-        final graph = Graph<String, void>.create(isDirected: true);
+        final graph = Graph<String, void>(isDirected: true);
         graph.addEdge('Hello', 'World');
         graph.removeEdge('Hello', 'World');
         expect(graph.vertices, unorderedEquals(['Hello', 'World']));
@@ -80,7 +80,7 @@ void main() {
         expectInvariants(graph);
       });
       test('add edge, remove first vertex', () {
-        final graph = Graph<String, void>.create(isDirected: true);
+        final graph = Graph<String, void>(isDirected: true);
         graph.addEdge('Hello', 'World');
         graph.removeVertex('Hello');
         expect(graph.vertices, ['World']);
@@ -88,7 +88,7 @@ void main() {
         expectInvariants(graph);
       });
       test('add edge, remove second vertex', () {
-        final graph = Graph<String, void>.create(isDirected: true);
+        final graph = Graph<String, void>(isDirected: true);
         graph.addEdge('Hello', 'World');
         graph.removeVertex('World');
         expect(graph.vertices, ['Hello']);
@@ -97,7 +97,7 @@ void main() {
       });
     });
     group('querying', () {
-      final graph = Graph<int, String>.create(isDirected: true)
+      final graph = Graph<int, String>(isDirected: true)
         ..addEdge(0, 1, value: 'a')
         ..addEdge(1, 2, value: 'b');
       test('invariants', () {
@@ -171,41 +171,41 @@ void main() {
   });
   group('reversed', () {
     test('empty', () {
-      final graph = Graph<String, int>.create(isDirected: true).reversed;
+      final graph = Graph<String, int>(isDirected: true).reversed;
       expect(graph.isDirected, isTrue);
       expect(graph.vertices, isEmpty);
       expect(graph.edges, isEmpty);
       expectInvariants(graph);
     });
     test('reversing directed', () {
-      final graph = Graph<String, void>.create(isDirected: true);
+      final graph = Graph<String, void>(isDirected: true);
       graph.addEdge('a', 'b');
       final reversedGraph = graph.reversed;
       expect(reversedGraph.reversed, same(graph));
     });
     test('reversing undirected', () {
-      final graph = Graph<String, void>.create(isDirected: false);
+      final graph = Graph<String, void>(isDirected: false);
       graph.addEdge('a', 'b');
       final reversedGraph = graph.reversed;
       expect(reversedGraph, same(graph));
     });
     group('modifying', () {
       test('add vertex', () {
-        final graph = Graph<String, int>.create(isDirected: true).reversed;
+        final graph = Graph<String, int>(isDirected: true).reversed;
         graph.addVertex('Hello');
         expect(graph.vertices, ['Hello']);
         expect(graph.edges, isEmpty);
         expectInvariants(graph);
       });
       test('add vertices', () {
-        final graph = Graph<String, int>.create(isDirected: true).reversed;
+        final graph = Graph<String, int>(isDirected: true).reversed;
         graph.addVertices(['Hello', 'World']);
         expect(graph.vertices, ['Hello', 'World']);
         expect(graph.edges, isEmpty);
         expectInvariants(graph);
       });
       test('remove vertex', () {
-        final graph = Graph<String, int>.create(isDirected: true).reversed;
+        final graph = Graph<String, int>(isDirected: true).reversed;
         graph.addVertex('Hello');
         graph.removeVertex('Hello');
         expect(graph.vertices, isEmpty);
@@ -213,7 +213,7 @@ void main() {
         expectInvariants(graph);
       });
       test('add edge', () {
-        final graph = Graph<String, int>.create(isDirected: true).reversed;
+        final graph = Graph<String, int>(isDirected: true).reversed;
         graph.addEdge('Hello', 'World', value: 42);
         expect(graph.vertices, unorderedEquals(['Hello', 'World']));
         expect(graph.edges, [
@@ -223,7 +223,7 @@ void main() {
         expectInvariants(graph);
       });
       test('add self-edge', () {
-        final graph = Graph<String, int>.create(isDirected: true).reversed;
+        final graph = Graph<String, int>(isDirected: true).reversed;
         graph.addEdge('Myself', 'Myself', value: 42);
         expect(graph.vertices, unorderedEquals(['Myself']));
         expect(
@@ -235,9 +235,7 @@ void main() {
         expectInvariants(graph);
       });
       test('put edge', () {
-        final graph = Graph<String, List<int>>.create(
-          isDirected: true,
-        ).reversed;
+        final graph = Graph<String, List<int>>(isDirected: true).reversed;
         graph.putEdge('a', 'b', () => []).add(1);
         graph.putEdge('b', 'a', () => []).add(2);
         expect(graph.vertices, unorderedEquals(['a', 'b']));
@@ -251,7 +249,7 @@ void main() {
         expectInvariants(graph);
       });
       test('remove edge', () {
-        final graph = Graph<String, void>.create(isDirected: true).reversed;
+        final graph = Graph<String, void>(isDirected: true).reversed;
         graph.addEdge('Hello', 'World');
         graph.removeEdge('Hello', 'World');
         expect(graph.vertices, unorderedEquals(['Hello', 'World']));
@@ -259,7 +257,7 @@ void main() {
         expectInvariants(graph);
       });
       test('add edge, remove first vertex', () {
-        final graph = Graph<String, void>.create(isDirected: true).reversed;
+        final graph = Graph<String, void>(isDirected: true).reversed;
         graph.addEdge('Hello', 'World');
         graph.removeVertex('Hello');
         expect(graph.vertices, ['World']);
@@ -267,7 +265,7 @@ void main() {
         expectInvariants(graph);
       });
       test('add edge, remove second vertex', () {
-        final graph = Graph<String, void>.create(isDirected: true).reversed;
+        final graph = Graph<String, void>(isDirected: true).reversed;
         graph.addEdge('Hello', 'World');
         graph.removeVertex('World');
         expect(graph.vertices, ['Hello']);
@@ -276,7 +274,7 @@ void main() {
       });
     });
     group('querying', () {
-      final directed = Graph<int, String>.create(isDirected: true)
+      final directed = Graph<int, String>(isDirected: true)
         ..addEdge(0, 1, value: 'a')
         ..addEdge(1, 2, value: 'b');
       final graph = directed.reversed;
@@ -351,28 +349,28 @@ void main() {
   });
   group('undirected', () {
     test('empty', () {
-      final graph = Graph<String, int>.create(isDirected: false);
+      final graph = Graph<String, int>(isDirected: false);
       expect(graph.vertices, isEmpty);
       expect(graph.edges, isEmpty);
       expectInvariants(graph);
     });
     group('modifying', () {
       test('add vertex', () {
-        final graph = Graph<String, int>.create(isDirected: false);
+        final graph = Graph<String, int>(isDirected: false);
         graph.addVertex('Hello');
         expect(graph.vertices, ['Hello']);
         expect(graph.edges, isEmpty);
         expectInvariants(graph);
       });
       test('add vertices', () {
-        final graph = Graph<String, int>.create(isDirected: false);
+        final graph = Graph<String, int>(isDirected: false);
         graph.addVertices(['Hello', 'World']);
         expect(graph.vertices, ['Hello', 'World']);
         expect(graph.edges, isEmpty);
         expectInvariants(graph);
       });
       test('remove vertex', () {
-        final graph = Graph<String, int>.create(isDirected: false);
+        final graph = Graph<String, int>(isDirected: false);
         graph.addVertex('Hello');
         graph.removeVertex('Hello');
         expect(graph.vertices, isEmpty);
@@ -380,7 +378,7 @@ void main() {
         expectInvariants(graph);
       });
       test('add edge', () {
-        final graph = Graph<String, int>.create(isDirected: false);
+        final graph = Graph<String, int>(isDirected: false);
         graph.addEdge('Hello', 'World', value: 42);
         expect(graph.vertices, unorderedEquals(['Hello', 'World']));
         expect(
@@ -393,7 +391,7 @@ void main() {
         expectInvariants(graph);
       });
       test('add self-edge', () {
-        final graph = Graph<String, int>.create(isDirected: false);
+        final graph = Graph<String, int>(isDirected: false);
         graph.addEdge('Myself', 'Myself', value: 42);
         expect(graph.vertices, unorderedEquals(['Myself']));
         expect(
@@ -405,7 +403,7 @@ void main() {
         expectInvariants(graph);
       });
       test('put edge', () {
-        final graph = Graph<String, List<int>>.create(isDirected: false);
+        final graph = Graph<String, List<int>>(isDirected: false);
         graph.putEdge('a', 'b', () => []).add(1);
         graph.putEdge('b', 'a', () => []).add(2);
         expect(graph.vertices, unorderedEquals(['a', 'b']));
@@ -419,7 +417,7 @@ void main() {
         expectInvariants(graph);
       });
       test('remove edge', () {
-        final graph = Graph<String, void>.create(isDirected: false);
+        final graph = Graph<String, void>(isDirected: false);
         graph.addEdge('Hello', 'World');
         graph.removeEdge('Hello', 'World');
         expect(graph.vertices, unorderedEquals(['Hello', 'World']));
@@ -427,7 +425,7 @@ void main() {
         expectInvariants(graph);
       });
       test('add edge, remove first vertex', () {
-        final graph = Graph<String, void>.create(isDirected: false);
+        final graph = Graph<String, void>(isDirected: false);
         graph.addEdge('Hello', 'World');
         graph.removeVertex('Hello');
         expect(graph.vertices, ['World']);
@@ -435,7 +433,7 @@ void main() {
         expectInvariants(graph);
       });
       test('add edge, remove second vertex', () {
-        final graph = Graph<String, void>.create(isDirected: false);
+        final graph = Graph<String, void>(isDirected: false);
         graph.addEdge('Hello', 'World');
         graph.removeVertex('World');
         expect(graph.vertices, ['Hello']);
@@ -444,7 +442,7 @@ void main() {
       });
     });
     group('querying', () {
-      final graph = Graph<int, String>.create(isDirected: false)
+      final graph = Graph<int, String>(isDirected: false)
         ..addEdge(0, 1, value: 'a')
         ..addEdge(1, 2, value: 'b');
       test('invariants', () {
@@ -540,13 +538,13 @@ void main() {
       });
     });
     test('reversed', () {
-      final graph = Graph<String, int>.create(isDirected: false);
+      final graph = Graph<String, int>(isDirected: false);
       expect(graph.reversed, same(graph));
     });
   });
   group('forwarding', () {
     test('directed', () {
-      final base = Graph<String, int>.create(isDirected: true);
+      final base = Graph<String, int>(isDirected: true);
       final graph = ForwardingGraph<String, int>(base);
       expect(graph.isDirected, isTrue);
       expect(graph.isUnmodifiable, isFalse);
@@ -576,7 +574,7 @@ void main() {
       graph.removeVertex('a');
     });
     test('undirected', () {
-      final base = Graph<String, int>.create(isDirected: false);
+      final base = Graph<String, int>(isDirected: false);
       final graph = ForwardingGraph<String, int>(base);
       expect(graph.isDirected, isFalse);
       expect(graph.isUnmodifiable, isFalse);
@@ -614,7 +612,7 @@ void main() {
   });
   group('unmodifiable', () {
     test('empty', () {
-      final graph = Graph<String, int>.create(isDirected: true);
+      final graph = Graph<String, int>(isDirected: true);
       expect(graph.isUnmodifiable, isFalse);
       final unmodifiable = graph.unmodifiable;
       expect(unmodifiable.isUnmodifiable, isTrue);
@@ -622,7 +620,7 @@ void main() {
       expectInvariants(unmodifiable);
     });
     test('delegates', () {
-      final graph = Graph<String, int>.create(isDirected: true);
+      final graph = Graph<String, int>(isDirected: true);
       graph.addEdge('a', 'b', value: 42);
       final unmodifiable = graph.unmodifiable;
       expect(unmodifiable.neighboursOf('a'), ['b']);
@@ -634,7 +632,7 @@ void main() {
       expect(unmodifiable.getEdge('a', 'b'), isEdge('a', 'b', value: 42));
     });
     test('errors', () {
-      final graph = Graph<String, void>.create(isDirected: true).unmodifiable;
+      final graph = Graph<String, void>(isDirected: true).unmodifiable;
       expect(() => graph.addVertex('a'), throwsUnsupportedError);
       expect(() => graph.addEdge('a', 'b'), throwsUnsupportedError);
       expect(() => graph.removeVertex('a'), throwsUnsupportedError);
@@ -644,7 +642,7 @@ void main() {
   });
   group('where', () {
     test('vertex predicate on directed graph', () {
-      final base = Graph<String, void>.create(isDirected: true);
+      final base = Graph<String, void>(isDirected: true);
       base
         ..addEdge('a', 'b')
         ..addEdge('b', 'c')
@@ -678,7 +676,7 @@ void main() {
       expect(graph.successorsOf('c'), ['a']);
     });
     test('vertex predicate on undirected graph', () {
-      final base = Graph<String, void>.create(isDirected: false);
+      final base = Graph<String, void>(isDirected: false);
       base
         ..addEdge('a', 'b')
         ..addEdge('b', 'c')
@@ -718,7 +716,7 @@ void main() {
       expect(graph.successorsOf('c'), ['a']);
     });
     test('vertex predicate filtering nothing', () {
-      final base = Graph<String, void>.create(isDirected: true);
+      final base = Graph<String, void>(isDirected: true);
       base
         ..addEdge('a', 'b')
         ..addEdge('b', 'c')
@@ -729,7 +727,7 @@ void main() {
       expect(graph.edges, unorderedEquals(base.edges));
     });
     test('vertex predicate filtering everything', () {
-      final base = Graph<String, void>.create(isDirected: true);
+      final base = Graph<String, void>(isDirected: true);
       base
         ..addEdge('a', 'b')
         ..addEdge('b', 'c')
@@ -740,7 +738,7 @@ void main() {
       expect(graph.edges, isEmpty);
     });
     test('edge predicate on directed graph', () {
-      final base = Graph<String, void>.create(isDirected: true);
+      final base = Graph<String, void>(isDirected: true);
       base
         ..addEdge('a', 'b')
         ..addEdge('b', 'c')
@@ -776,7 +774,7 @@ void main() {
       expect(graph.successorsOf('c'), ['a']);
     });
     test('edge predicate on undirected graph', () {
-      final base = Graph<String, void>.create(isDirected: false);
+      final base = Graph<String, void>(isDirected: false);
       base
         ..addEdge('a', 'b')
         ..addEdge('b', 'c')
@@ -818,7 +816,7 @@ void main() {
       expect(graph.successorsOf('c'), ['a']);
     });
     test('edge predicate filtering nothing', () {
-      final base = Graph<String, void>.create(isDirected: true);
+      final base = Graph<String, void>(isDirected: true);
       base
         ..addEdge('a', 'b')
         ..addEdge('b', 'c')
@@ -829,7 +827,7 @@ void main() {
       expect(graph.edges, unorderedEquals(base.edges));
     });
     test('edge predicate filtering everything', () {
-      final base = Graph<String, void>.create(isDirected: true);
+      final base = Graph<String, void>(isDirected: true);
       base
         ..addEdge('a', 'b')
         ..addEdge('b', 'c')
@@ -840,7 +838,7 @@ void main() {
       expect(graph.edges, isEmpty);
     });
     test('no predicates', () {
-      final base = Graph<int, int>.create(isDirected: true);
+      final base = Graph<int, int>(isDirected: true);
       final graph = base.where();
       expect(graph, same(base));
     });
@@ -848,70 +846,100 @@ void main() {
   group('edge', () {
     group('directed', () {
       const a = Edge<int, void>.directed(1, 2);
-      const b = Edge<int, String>.directed(2, 1, value: 'b');
-      const c = Edge<int, String>.directed(2, 1, value: 'c');
+      const b = Edge<int, String>.directed(1, 2, value: 'a');
+      const c = Edge<int, String>.directed(2, 1, value: 'a');
+      const d = Edge<int, String>.directed(1, 2, value: 'b');
       test('create (without data)', () {
         expect(a.source, 1);
         expect(a.target, 2);
+        expect(a.isDirected, isTrue);
       });
       test('create (with data)', () {
-        expect(c.source, 2);
-        expect(c.target, 1);
-        expect(c.value, 'c');
+        expect(b.source, 1);
+        expect(b.target, 2);
+        expect(b.value, 'a');
+        expect(b.isDirected, isTrue);
       });
       test('equals', () {
         expect(a == a, isTrue);
         expect(a == b, isFalse);
         expect(a == c, isFalse);
+        expect(a == d, isFalse);
         expect(b == a, isFalse);
         expect(b == b, isTrue);
-        expect(b == c, isTrue);
+        expect(b == c, isFalse);
+        expect(b == d, isFalse);
         expect(c == a, isFalse);
-        expect(c == b, isTrue);
+        expect(c == b, isFalse);
         expect(c == c, isTrue);
+        expect(c == d, isFalse);
+        expect(d == a, isFalse);
+        expect(d == b, isFalse);
+        expect(d == c, isFalse);
+        expect(d == d, isTrue);
       });
       test('hashCode', () {
-        expect(a.hashCode == b.hashCode, isFalse);
-        expect(b.hashCode == c.hashCode, isTrue);
+        expect(a.hashCode, isNot(equals(b.hashCode)));
+        expect(a.hashCode, isNot(equals(c.hashCode)));
+        expect(a.hashCode, isNot(equals(d.hashCode)));
+        expect(b.hashCode, isNot(equals(c.hashCode)));
+        expect(b.hashCode, isNot(equals(d.hashCode)));
+        expect(c.hashCode, isNot(equals(d.hashCode)));
       });
       test('toString', () {
         expect(a.toString(), endsWith('(1 → 2)'));
-        expect(b.toString(), endsWith('(2 → 1, value: b)'));
-        expect(c.toString(), endsWith('(2 → 1, value: c)'));
+        expect(b.toString(), endsWith('(1 → 2, value: a)'));
+        expect(c.toString(), endsWith('(2 → 1, value: a)'));
+        expect(d.toString(), endsWith('(1 → 2, value: b)'));
       });
     });
     group('undirected', () {
       const a = Edge<int, void>.undirected(1, 2);
-      const b = Edge<int, String>.undirected(2, 1, value: 'b');
-      const c = Edge<int, String>.undirected(2, 3, value: 'c');
+      const b = Edge<int, String>.undirected(1, 2, value: 'a');
+      const c = Edge<int, String>.undirected(2, 1, value: 'a');
+      const d = Edge<int, String>.undirected(1, 2, value: 'b');
       test('create (without data)', () {
         expect(a.source, 1);
         expect(a.target, 2);
+        expect(a.isDirected, isFalse);
       });
       test('create (with data)', () {
-        expect(c.source, 2);
-        expect(c.target, 3);
-        expect(c.value, 'c');
+        expect(b.source, 1);
+        expect(b.target, 2);
+        expect(b.value, 'a');
+        expect(b.isDirected, isFalse);
       });
       test('equals', () {
         expect(a == a, isTrue);
-        expect(a == b, isTrue);
+        expect(a == b, isFalse);
         expect(a == c, isFalse);
-        expect(b == a, isTrue);
+        expect(a == d, isFalse);
+        expect(b == a, isFalse);
         expect(b == b, isTrue);
-        expect(b == c, isFalse);
+        expect(b == c, isTrue);
+        expect(b == d, isFalse);
         expect(c == a, isFalse);
-        expect(c == b, isFalse);
+        expect(c == b, isTrue);
         expect(c == c, isTrue);
+        expect(c == d, isFalse);
+        expect(d == a, isFalse);
+        expect(d == b, isFalse);
+        expect(d == c, isFalse);
+        expect(d == d, isTrue);
       });
       test('hashCode', () {
-        expect(a.hashCode == b.hashCode, isTrue);
-        expect(a.hashCode == c.hashCode, isFalse);
+        expect(a.hashCode, isNot(equals(b.hashCode)));
+        expect(a.hashCode, isNot(equals(c.hashCode)));
+        expect(a.hashCode, isNot(equals(d.hashCode)));
+        expect(b.hashCode, equals(c.hashCode));
+        expect(b.hashCode, isNot(equals(d.hashCode)));
+        expect(c.hashCode, isNot(equals(d.hashCode)));
       });
       test('toString', () {
         expect(a.toString(), endsWith('(1 — 2)'));
-        expect(b.toString(), endsWith('(2 — 1, value: b)'));
-        expect(c.toString(), endsWith('(2 — 3, value: c)'));
+        expect(b.toString(), endsWith('(1 — 2, value: a)'));
+        expect(c.toString(), endsWith('(2 — 1, value: a)'));
+        expect(d.toString(), endsWith('(1 — 2, value: b)'));
       });
     });
   });